src/test/java/ru/netology/NetologyJavaHW10/services/FilmManagerTest.java
package ru.netology.NetologyJavaHW10.services;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class FilmManagerTest {

    FilmManager manager = new FilmManager();

    FilmManager film1 = new FilmManager("Бладшот");
    FilmManager film2 = new FilmManager("Вперёд");
    FilmManager film3 = new FilmManager("Отель Белград");
    FilmManager film4 = new FilmManager("Джентльмены");
    FilmManager film5 = new FilmManager("Человек-невидимка");
    FilmManager film6 = new FilmManager("Тролли. Мировой тур");
    FilmManager film7 = new FilmManager("Номер один");


    @Test
    public void shouldAddFilms() {

        manager.addFilm(film1);


        String[] expected = {};
        String[] actual = manager.findAll();

        Assertions.assertArrayEquals(expected, actual);

    }

    @Test
    public void shouldAddOneFilm() {

        manager.addFilm("Бладшот");

        String[] expected = {"Бладшот"};
        String[] actual = manager.findAll();

        Assertions.assertArrayEquals(expected, actual);
    }

    @Test
    public void shouldAddSixFilms() {

        manager.addFilm("Бладшот");
        manager.addFilm("Вперёд");
        manager.addFilm("Отель Белград");
        manager.addFilm("Джентльмены");
        manager.addFilm("Человек-невидимка");
        manager.addFilm("Тролли. Мировой тур");

        String[] expected = { "Бладшот" };
        String[] actual = manager.findAll();

        Assertions.assertArrayEquals(expected, actual);
    }

    @Test

    public void shouldAddAllFilms() {

        manager.addFilm("Бладшот");
        manager.addFilm("Вперёд");
        manager.addFilm("Отель Белград");
        manager.addFilm("Джентльмены");
        manager.addFilm("Человек-невидимка");
        manager.addFilm("Тролли. Мировой тур");
        manager.addFilm("Номер один");

        String[] expected = {};
        String[] actual = manager.findAll();

        Assertions.assertArrayEquals(expected, actual);

    }

    @Test

    public void shouldShowLastFilmsReversed() {

        manager.addFilm("Отель Белград");
        manager.addFilm("Джентльмены");
        manager.addFilm("Человек-невидимка");
        manager.addFilm("Тролли. Мировой тур");
        manager.addFilm("Номер один");

        String[] expected = {};
        String[] actual = manager.findLast();

        Assertions.assertArrayEquals(expected, actual);
    }


}
